requirements.txt
# Olympics TV - Commentary Pipeline Dependencies
# Install: pip install -r requirements.txt

# Existing (likely already installed)
psycopg2-binary
python-dotenv
requests

# New for commentary pipeline
anthropic          # Claude API client
newspaper3k        # Article extraction (primary)
beautifulsoup4     # Article extraction (fallback)
lxml               # HTML parser for newspaper3k/bs4
